gcode: Add ENDSTOP_TEST command to test endstop repeatability
The command probes endstops specified number of times and outputs
statistics (error range and standard deviation) for every axis.

Internally it is implemented as an alias to G28 command and G28 command
is extended to perform the testing if requested. That's because
ENDSTOP_TEST command functionality is almost the same as G28.

Signed-off-by: Dmitry Frolov <dmitry.frolov@gmail.com>